crunch: use _FILE_OFFSET_BITS 64 with fopen/fseek/ftell outside of Linux
It seems to work on macOS and FreeBSD without requiring
specific #ifdef for any them. It also works on Linux but
we may still prefer *64 variants to avoid ambiguity.

Note: If needed, minizip/ioapi.h is known to provide more
compatibility code for old MSVC too, but what we have looks
good enough on that part.